+++ src/test/test_save_votes.cpp
@@ -36,13 +36,18 @@ void saveWhenNoVotesRemain() {
 
 } // namespace
 
+auto run_tests(char* argv[]) -> int {
+	RUN_TEST_IF_ARGUMENT_EQUALS(saveWhenSomeVotesRemain);
+	RUN_TEST_IF_ARGUMENT_EQUALS(saveWhenNoVotesRemain);
+	return 1;
+}
+
 int main(int argc, char* argv[]) {
 	ASSERT_EQ(argc, 2);
 	std::filesystem::remove(kTestFileName);
-
-	RUN_TEST_IF_ARGUMENT_EQUALS(saveWhenSomeVotesRemain);
-	RUN_TEST_IF_ARGUMENT_EQUALS(saveWhenNoVotesRemain);
-
+	if (run_tests(argv) != 0) {
+		return 1;
+	}
 	std::filesystem::remove(kTestFileName);
-	return 1;
+	return 0;
 }
